Mock Nordvpn commands with persistent state
When logging in, out, connecting or disconnecting, the mock
commands class for nordvpn saves the state in instance variables
to then act as logged in, logged out, connected, or disconnected.

This makes it mock the real nordvpn commands better, because the real
nordvpn commands have some effect on the nordvpn shell application and
thus they are later reflected when getting the status or account info.